index and term use unsigned long instead of int/uint

+++ include/raft_log.h
@@ -22,11 +22,11 @@ int log_append_entry(log_t* me_, raft_entry_t* c);
 
 /**
  * @return number of entries held within log */
-int log_count(log_t* me_);
+unsigned long log_count(log_t* me_);
 
 /**
  * Delete all logs from this log onwards */
-int log_delete(log_t* me_, int idx);
+int log_delete(log_t* me_, unsigned long idx);
 
 /**
  * Empty the queue. */
@@ -39,18 +39,18 @@ int log_poll(log_t * me_, void** etyp);
 /** Get an array of entries from this index onwards.
  * This is used for batching.
  */
-raft_entry_t* log_get_from_idx(log_t* me_, int idx, int *n_etys);
+raft_entry_t* log_get_from_idx(log_t* me_, unsigned long idx, unsigned long *n_etys);
 
-raft_entry_t* log_get_at_idx(log_t* me_, int idx);
+raft_entry_t* log_get_at_idx(log_t* me_, unsigned long idx);
 
 /**
  * @return youngest entry */
 raft_entry_t *log_peektail(log_t * me_);
 
-int log_get_current_idx(log_t* me_);
+unsigned long log_get_current_idx(log_t* me_);
 
-int log_load_from_snapshot(log_t *me_, int idx, int term);
+int log_load_from_snapshot(log_t *me_, unsigned long idx, unsigned long term);
 
-int log_get_base(log_t* me_);
+unsigned long log_get_base(log_t* me_);
 
 #endif /* RAFT_LOG_H_ */
